requirements.txt
anyio==3.6.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.8.0
gevent==22.10.2
greenlet==2.0.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
kraky==2022.6.15
numpy==1.23.4
packaging==21.3
pandas==1.5.2
platformdirs==2.5.4
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tox==3.27.1
typer==0.4.2
urllib3==1.26.12
virtualenv==20.16.7
websocket-client==1.4.2
zope.event==4.5.0
zope.interface==5.5.1
